refactor files and components to reorganize screen
