+++ frontend/components/Camera/CameraScreen.jsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import CameraContoller from './Camera';
+
+function CameraScreen() {
+  return (<CameraContoller />);
+}
+
+export default CameraScreen;
