+++ frontend/components/Plants/CardItem.jsx
